requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
accelerate
pymupdf
numpy
